% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountGeneMutations.R
\name{count_gene_mutations}
\alias{count_gene_mutations}
\title{Count gene mutations}
\usage{
count_gene_mutations(
  table,
  gene.col,
  gt.col,
  flag = c("Alleles", "Snps", "Carrier")
)
}
\arguments{
\item{table}{A table containing the information of each and every gene the samples have been tested on.}

\item{gene.col}{The number of the gene column when counting from the left.}

\item{gt.col}{The number of the first sample column when counting from the left.}

\item{flag}{"Allele", "Snps", or "Carrier", indicating the table to be generated}
}
\value{
flag = Alleles, A table containing the number of mutated alleles for each gene in each sample.
flag = Snps, A table containing the number of mutated locus in each gene for each sample.
flag = Carrier, A table indicating the presence of mutation in each gene for each sample.
}
\description{
Count gene mutations
}
\examples{
count_gene_mutations(table, 2, 3, "Alleles")
}
