% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CountGeneSetVariant.R
\name{count_gene_set_variant}
\alias{count_gene_set_variant}
\title{Count gene set variant}
\usage{
count_gene_set_variant(Table, GeneSet, Name)
}
\arguments{
\item{Table}{A tibble indicating if each individual has mutation in each of the genes}

\item{GeneSet}{A one-col table containing the genes to be observed}

\item{Name}{A string, the name of the gene set, e.g. "high_expressed_genes"}
}
\value{
count_gene_set_variant() returns a table specifying the number of mutations in the genes observed for each and every individual
}
\description{
Count gene set variant
}
\examples{
count_gene_set_variant(table, Set, "high_expressed_genes")
}
