% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vars_in_gene_set_per_sample.R
\name{count_vars_in_gene_set_per_sample}
\alias{count_vars_in_gene_set_per_sample}
\title{count genotype of specified genes}
\usage{
count_vars_in_gene_set_per_sample(table, gene.set, gene.set.name)
}
\arguments{
\item{Table}{a dataframe contains a colume of gene names and columes of population genetype which have been processed into numeric form.}

\item{GeneSet}{The name of gene(can be more than one) you are interested in.}

\item{Name}{the name of geneset you come up with.}
}
\value{
The return of the function is a table containing two columes named by "Sample ID" and user's input "Name". The "Name" colume dislpays the sum of genotypes of the genes offered by the user.
}
\description{
the function is used to count the genotype of specified genes inputed by the user. A reference of which geneset has a higher impact on the disease.
}
