% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vars_per_gene_per_sample.R
\name{count_vars_per_gene_per_sample}
\alias{count_vars_per_gene_per_sample}
\title{Count gene mutations}
\usage{
count_vars_per_gene_per_sample(
  table,
  gene.col,
  gt.col,
  type = c("allele", "locus", "carrier")
)
}
\arguments{
\item{table}{A table containing the information of each and every gene the samples have been tested on.}

\item{gene.col}{The number of the gene column when counting from the left.}

\item{gt.col}{The number of the first sample column when counting from the left.}

\item{type}{"Allele", "Snps", or "Carrier", indicating the table to be generated}
}
\value{
type = Alleles, A table containing the number of mutated alleles for each gene in each sample.
type = Snps, A table containing the number of mutated locus in each gene for each sample.
type = Carrier, A table indicating the presence of mutation in each gene for each sample.
}
\description{
Count gene mutations
}
\examples{
\dontrun{
count_vars_per_gene_per_sample(table = variants, gene.col = 8, gt.col = 12, type = "allele")
}
}
