% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_for_multiple_tests.R
\name{glm_for_multiple_tests}
\alias{glm_for_multiple_tests}
\title{Use modeling to calculate correlation}
\usage{
glm_for_multiple_tests(
  ped,
  outcome,
  interested.variables,
  covariates,
  tidy.output = FALSE,
  conf.int = FALSE,
  conf.level = 0.95,
  exponentiate = FALSE
)
}
\arguments{
\item{ped}{A raw dataframe from the bioinformation database.}

\item{outcome}{The type of outcome to be tested.}

\item{covariates}{The covariates(e.g. age, gender)help testing the correlation.}

\item{geneset}{The genesets(can be more than one)that users want to test the correlation with the outcome.}
}
\value{
The return of this function is a table with p-value, conf.high and conf.high that give users direction on which geneset has greatest correlation to the disease.
}
\description{
With covariants and genesets, this function would calculate the necessary values that uncovers the correlation between genesets and outcomes.
}
