% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestGeneSetRelevancy.R
\name{test_gene_set_relevancy}
\alias{test_gene_set_relevancy}
\title{Test gene set relevancy}
\usage{
test_gene_set_relevancy(rawTable, GeneSet, CVnames, Trait, conf.int = FALSE)
}
\arguments{
\item{rawTable}{A PED format table containing the phenotypes and covariates of each sample.}

\item{GeneSet}{A one column table containing the col names in the PED table for each gene in the gene set.}

\item{CVnames}{A one column table containing the col names in the PED table for each covariate in the gene set.}

\item{Trait}{A character specifying the phenotype to be investigated.}

\item{conf.int}{\itemize{
\item 
}}
}
\value{
A tidy table including regular summary from the glm, e.g. beta, std.error, p.val, odds.ratio, ...
}
\description{
Test gene set relevancy
}
\examples{
test_gene_set_relevancy(file.path(R.home("raw_table")), file.path(R.home("gene_set")), file.path(R.home("covariates")), DISEASE)
}
