% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vars_in_gene_set_per_sample.R
\name{count_vars_in_gene_set_per_sample}
\alias{count_vars_in_gene_set_per_sample}
\title{Count the total number of counting metric in a gene set for each sample}
\usage{
count_vars_in_gene_set_per_sample(table, gene.set, gene.set.name)
}
\arguments{
\item{table}{A tibble which has first column representing gene name and the rest of columns representing the counting metrics of samples. An example is the output of the function count_vars_per_gene_per_sample.}

\item{gene.set}{A vector of gene names specifying a gene set. Usually it is a subset of genes from the first column of input table.}

\item{gene.set.name}{A string specifying the name of the gene set which will be used as the column name in the output}
}
\value{
A tibble with two columns. The first column "sampleID" represents sample names and the second column named with \emph{gene.set.name} represents the total number of metric in the input \emph{gene.set}.
}
\description{
This function count the total number of counting metric (i.e. alleles, locus or carrier-status from function count_vars_per_gene_per_sample) in a gene set for each sample.
}
\examples{
\dontrun{
data("1000g")

# generate allele count table for each gene and each sample
gene.ac <- count_vars_per_gene_per_sample(table = variants, gene.col = 8, gt.col = 12, type = "allele")

# here just pick some genes and use as an example as a gene set input
gene.list <- gene.ac$GENE[10:14]

# count the total number of alleles in the gene set for each sample
geneset.ac <- count_vars_in_gene_set_per_sample(table = gene.ac, gene.set = gene.list, gene.set.name = "genesetA")
}
}
