% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_vars_per_gene_per_sample.R
\name{count_vars_per_gene_per_sample}
\alias{count_vars_per_gene_per_sample}
\title{Generate the metric table per gene per sample}
\usage{
count_vars_per_gene_per_sample(
  table,
  gene.col,
  gt.col,
  type = c("allele", "locus", "carrier")
)
}
\arguments{
\item{table}{A data frame which contains information of variants first and then followed by genotypes of all samples. The information of variants, annotated gene and sample names are required; Chrom, pos, ID, ref, alt, effect, impact, etc are optional.}

\item{gene.col}{The position of column representing annotated gene.}

\item{gt.col}{The position of column representing the genotype of the first sample. By default, all following columns after gt.col are genotypes of the rest samples.}

\item{type}{Control which type of counting metric to use.
\itemize{
\item "allele" count the total number of allele count per gene per sample
\item "locus" count the total number of genetic positions with mutation per gene per sample
\item "carrier", return the status of the sample carrying at least one mutation for each gene and each sample. 1 for carrier, 0 for non-carrier.
}}
}
\value{
A data frame which contains the count of allele/locus/carrier-status per gene per sample. The first column represents annotated gene, and the rest of columns represent the count of allele/locus/carrier-status of all samples, depending on which the parameter of \emph{type} you choose
}
\description{
This function count the total number of alleles, the total number of locus, or the status of carrying any mutations per gene per sample depending on which type of counting metric you choose.
}
\examples{
\dontrun{
data("1000g")

# generate allele count table for each gene and each sample
gene.ac <- count_vars_per_gene_per_sample(table = variants, gene.col = 8, gt.col = 12, type = "allele")

# generate locus count table for each gene and each sample
gene.locus <- count_vars_per_gene_per_sample(table = variants, gene.col = 8, gt.col = 12, type = "locus")

# generate carrier status table for each gene and each sample
gene.carrier <- count_vars_per_gene_per_sample(table = variants, gene.col = 8, gt.col = 12, type = "carrier")

}
}
